Better fix for MSVC's PMOVSX/ZX bug, and have intrinsics do full inlining by default in all build targets (makes them behave more like the real intrinsics).
Turns out the bug does show up in VS2010 as well as VS2011 beta -- and that VS2011 did *not* affect my VS2010 install.  It is just very circumstantial as to which build+optimization options the bug shows in, and the project I was using on VS2010 had customized debug/devel/release target settings that differed a bit from my VS2011 beta project.